+++ packages/agentation-vue/src/constants.ts
@@ -1,2 +1,3 @@
 export const VA_DATA_ATTR = 'data-agentation-vue'
 export const VA_DATA_ATTR_SELECTOR = '[data-agentation-vue]'
+export const VA_VERSION = '0.2.10'